refactor(dto): split runner contracts into request and response
