+++ core/src/main/java/com/marmitt/core/application/usecase/runner/QueryRunnerUseCase.java
@@ -1,7 +1,7 @@
 package com.marmitt.core.application.usecase.runner;
 
 import com.marmitt.core.dto.portfolio.response.TransactionDto;
-import com.marmitt.core.dto.runner.RunnerDto;
+import com.marmitt.core.dto.runner.response.RunnerDto;
 import com.marmitt.core.enums.TransactionStatus;
 import com.marmitt.core.ports.inbound.runner.QueryRunnerPort;
 import com.marmitt.core.ports.outbound.repository.StrategyRunnerRepositoryPort;
@@ -50,4 +50,5 @@ public class QueryRunnerUseCase implements QueryRunnerPort {
                 .toList();
     }
 }
+
