+++ core/src/main/java/com/marmitt/core/ports/inbound/runner/CreateRunnerPort.java
@@ -1,9 +1,10 @@
 package com.marmitt.core.ports.inbound.runner;
 
-import com.marmitt.core.dto.runner.CreateRunnerRequest;
-import com.marmitt.core.dto.runner.CreateRunnerResponse;
+import com.marmitt.core.dto.runner.request.CreateRunnerRequest;
+import com.marmitt.core.dto.runner.response.CreateRunnerResponse;
 
 public interface CreateRunnerPort {
 
     CreateRunnerResponse execute(CreateRunnerRequest request);
 }
+
