+++ core/src/main/java/com/marmitt/core/ports/inbound/runner/QueryRunnerPort.java
@@ -1,7 +1,7 @@
 package com.marmitt.core.ports.inbound.runner;
 
 import com.marmitt.core.dto.portfolio.response.TransactionDto;
-import com.marmitt.core.dto.runner.RunnerDto;
+import com.marmitt.core.dto.runner.response.RunnerDto;
 
 import java.util.List;
 import java.util.UUID;
@@ -12,4 +12,5 @@ public interface QueryRunnerPort {
 
     List<TransactionDto> findTransactionsByRunnerId(UUID runnerId);
 }
+
