+++ spring-application/src/test/java/com/marmitt/application/spring/bootstrap/ConcurrentSellLockIntegrationTest.java
@@ -7,8 +7,8 @@ import com.marmitt.core.domain.runner.StrategyRunner;
 import com.marmitt.core.domain.runner.Transaction;
 import com.marmitt.core.dto.portfolio.request.CreatePortfolioRequest;
 import com.marmitt.core.dto.portfolio.response.CreatePortfolioResponse;
-import com.marmitt.core.dto.runner.CreateRunnerRequest;
-import com.marmitt.core.dto.runner.CreateRunnerResponse;
+import com.marmitt.core.dto.runner.request.CreateRunnerRequest;
+import com.marmitt.core.dto.runner.response.CreateRunnerResponse;
 import com.marmitt.core.enums.PositionStatus;
 import com.marmitt.core.enums.TransactionStatus;
 import com.marmitt.core.enums.TransactionType;
@@ -331,4 +331,5 @@ class ConcurrentSellLockIntegrationTest {
     private record LockResult(UUID transactionId, boolean locked) {
     }
 }
+
