+++ spring-application/src/test/java/com/marmitt/application/spring/bootstrap/MockOrderOverrideIntegrationTestSupport.java
@@ -7,8 +7,8 @@ import com.marmitt.core.domain.runner.StrategyRunner;
 import com.marmitt.core.domain.runner.Transaction;
 import com.marmitt.core.dto.portfolio.request.CreatePortfolioRequest;
 import com.marmitt.core.dto.portfolio.response.CreatePortfolioResponse;
-import com.marmitt.core.dto.runner.CreateRunnerRequest;
-import com.marmitt.core.dto.runner.CreateRunnerResponse;
+import com.marmitt.core.dto.runner.request.CreateRunnerRequest;
+import com.marmitt.core.dto.runner.response.CreateRunnerResponse;
 import com.marmitt.core.dto.websocket.data.OrderDataDto;
 import com.marmitt.core.dto.websocket.request.SendOrderRequest;
 import com.marmitt.core.enums.OrderSide;
@@ -601,4 +601,5 @@ abstract class MockOrderOverrideIntegrationTestSupport {
     protected record BalanceRow(BigDecimal available, BigDecimal reserved, BigDecimal realized) {
     }
 }
+
