+++ spring-application/src/test/java/com/marmitt/application/spring/bootstrap/OrderTerminationConciliationIntegrationTest.java
@@ -8,8 +8,8 @@ import com.marmitt.core.domain.runner.Position;
 import com.marmitt.core.domain.runner.Transaction;
 import com.marmitt.core.dto.portfolio.request.CreatePortfolioRequest;
 import com.marmitt.core.dto.portfolio.response.CreatePortfolioResponse;
-import com.marmitt.core.dto.runner.CreateRunnerRequest;
-import com.marmitt.core.dto.runner.CreateRunnerResponse;
+import com.marmitt.core.dto.runner.request.CreateRunnerRequest;
+import com.marmitt.core.dto.runner.response.CreateRunnerResponse;
 import com.marmitt.core.dto.websocket.data.OrderDataDto;
 import com.marmitt.core.enums.PositionStatus;
 import com.marmitt.core.enums.TransactionStatus;
@@ -392,4 +392,5 @@ class OrderTerminationConciliationIntegrationTest {
         }
     }
 }
+
